Remove the Context interface
Curveball had a Context interface and BaseContext class. The idea was
that it might be possible for users to create alternative
implementations.

In practice this has never happened, and is probably not that useful.
The idea of using interfaces for everything is also something that I
blindly applied before fully getting my bearings in Typescript.

There's some drawbacks to the interface. Extending the Context by other
middlewares such as the router, oauth2 middleware and others is a really
useful feature. Every time somebody wants to extend it, they are forced
to extend both the interface and the class.

This is a breaking change so this will be part of Curveball 0.18.
