+++ src/index.ts
@@ -1,6 +1,5 @@
 import { default as Application, invokeMiddlewares, Middleware, middlewareCall } from './application';
-import BaseContext from './base-context';
-export { Context, WsContext } from './context';
+export { Context } from './context';
 import Headers from './headers';
 import MemoryRequest from './memory-request';
 import MemoryResponse from './memory-response';
@@ -11,7 +10,6 @@ import { conditionalCheck } from './conditional';
 export default Application;
 export {
   Application,
-  BaseContext,
   conditionalCheck,
   Headers,
   invokeMiddlewares,
